Give runners the instance id as name
By default the hostname is used. That looks like ip-10-136-84-150. That
is the instances public ip4 address. If the instances cycle fast enough,
EC2 re-uses they same IP and we get clashing names on the github side
and the instance registration fails. By using the instance id, we ensure
the name in github is always unique and can match it to the instance
easily.