Stop all started instances
Remove all started AWS instances. Updates the instance id input and
output to support a comma seperated list of id.

Remove all runners with the same label, that is the unique label
generated for this set of runners. Means we clean up fully when using
runner-count > 1 and don't leave disconnected runners behind in GitHub.

Update dist