remove default ec2VolumeSize or ec2VolumeType settings
Prior to this change, if a user did not specify ec2-volume-size or
ec2-volume-type, we would send default values of "8" and "gp2" to the EC2
API.

This was a problem if the user's AMI used a snapshot with other settings
(like a bigger disk size, or another disk type like "io1"). AWS could
fail to start the nodes if the AMI was incompatible with these defaults.

Update the action to only send ec2VolumeSize or ec2VolumeType in the API
call if the user has specified a value (not an empty string).

With this change, we reintroduce the behavior prior to
8b37f736c69ba6af391e437447d3c07548478d78.

This change preserves the default for ec2-device-name, since we need
that default value in order to construct BlockDeviceSettings.